feat: implement parallel test execution for Playwright tests
- Separate global operations into session-management-global.spec.ts
- Add parallel configuration to test files that create isolated sessions
- Update playwright.config.ts with parallel execution settings
- Create test groups: chromium-parallel and chromium-serial
- Add test execution script and comprehensive documentation
- Tests can now run concurrently, significantly reducing execution time
