+++ web/PARALLEL_TESTING_CHANGES.md
@@ -1,0 +1,79 @@
+# Parallel Testing Implementation Summary
+
+## Changes Made
+
+### 1. Created `parallel-playwright` branch
+- New branch created from main for all parallel testing improvements
+
+### 2. Separated Global Tests
+- Created `session-management-global.spec.ts` for tests that affect all sessions
+- Moved "Kill All" and "Filter Sessions" tests from `session-management-advanced.spec.ts`
+- These tests must run serially to avoid interfering with other tests
+
+### 3. Added Parallel Configuration
+Added `test.describe.configure({ mode: 'parallel' });` to these test files:
+- `session-creation.spec.ts`
+- `basic-session.spec.ts` 
+- `minimal-session.spec.ts`
+- `debug-session.spec.ts`
+- `ui-features.spec.ts`
+- `test-session-persistence.spec.ts`
+- `session-navigation.spec.ts`
+- `session-management.spec.ts`
+- `session-management-advanced.spec.ts`
+
+### 4. Updated playwright.config.ts
+- Changed `fullyParallel: true` (was false)
+- Changed `workers: process.env.CI ? 2 : undefined` (was 1)
+- Created two project configurations:
+  - `chromium-parallel` - for tests that can run concurrently
+  - `chromium-serial` - for tests that must run sequentially
+
+### 5. Created Helper Script
+- `scripts/test-parallel.sh` - Easy way to run different test configurations
+- Options: all, parallel, serial, debug, ui
+
+### 6. Documentation
+- `src/test/playwright/PARALLEL_TESTING.md` - Comprehensive guide
+- Explains test organization, best practices, and troubleshooting
+
+## Benefits
+
+1. **Performance**: Tests can now run concurrently, significantly reducing total execution time
+2. **Isolation**: Better test organization ensures tests don't interfere with each other
+3. **Flexibility**: Can run parallel or serial tests separately as needed
+4. **Scalability**: Easy to add more workers or implement sharding in the future
+
+## Test Organization
+
+**Parallel Tests (31 tests)**
+- Session creation and management (individual operations)
+- UI features and navigation
+- Session persistence
+- Isolated session operations
+
+**Serial Tests (21 tests)**
+- Global operations (Kill All)
+- Session filtering with global state changes
+- Keyboard shortcuts (some may affect global state)
+- Terminal interaction (currently skipped)
+
+## Running Tests
+
+```bash
+# Run all tests
+./scripts/test-parallel.sh all
+
+# Run only parallel tests
+./scripts/test-parallel.sh parallel
+
+# Run only serial tests  
+./scripts/test-parallel.sh serial
+```
+
+## Next Steps
+
+1. Monitor test stability with parallel execution
+2. Adjust worker count based on performance
+3. Consider adding test sharding for even more parallelization
+4. Add performance metrics to track improvements
