+++ web/src/test/playwright/specs/basic-session.spec.ts
@@ -14,6 +14,9 @@ import { TestDataFactory } from '../utils/test-utils';
 // Use a unique prefix for this test suite
 const TEST_PREFIX = TestDataFactory.getTestSpecificPrefix('basic-session');
 
+// These tests create their own sessions and can run in parallel
+test.describe.configure({ mode: 'parallel' });
+
 test.describe('Basic Session Tests', () => {
   let sessionManager: TestSessionManager;
