+++ web/src/test/playwright/specs/debug-session.spec.ts
@@ -5,6 +5,9 @@ import { TestDataFactory } from '../utils/test-utils';
 // Use a unique prefix for this test suite
 const TEST_PREFIX = TestDataFactory.getTestSpecificPrefix('debug-session');
 
+// These tests create their own sessions and can run in parallel
+test.describe.configure({ mode: 'parallel' });
+
 test.describe('Debug Session Tests', () => {
   let sessionManager: TestSessionManager;
