+++ web/src/test/playwright/specs/session-navigation.spec.ts
@@ -4,6 +4,9 @@ import { takeDebugScreenshot } from '../helpers/screenshot.helper';
 import { createMultipleSessions } from '../helpers/session-lifecycle.helper';
 import { TestSessionManager } from '../helpers/test-data-manager.helper';
 
+// These tests create their own sessions and can run in parallel
+test.describe.configure({ mode: 'parallel' });
+
 test.describe('Session Navigation', () => {
   let sessionManager: TestSessionManager;
